+++ xyzpy/__init__.py
@@ -14,13 +14,14 @@ from .utils import (
 )
 from .gen.combo_runner import (
     combo_runner,
-    combo_runner_to_ds
+    combo_runner_to_ds,
 )
 from .gen.case_runner import (
     case_runner,
     find_union_coords,
     all_missing_ds,
     case_runner_to_ds,
+    case_runner_to_df,
     find_missing_cases,
     fill_missing_cases
 )
@@ -30,12 +31,16 @@ from .gen.batch import (
 )
 from .gen.farming import (
     Runner,
-    Harvester
+    Harvester,
+    label,
+    Sampler,
 )
 from .manage import (
     cache_to_disk,
     save_ds,
     load_ds,
+    save_df,
+    load_df,
     trimna,
     sort_dims,
     check_runs,
@@ -98,12 +103,15 @@ del get_versions
 __all__ = [
     "Runner",
     "Harvester",
+    "Sampler",
+    "label",
     "combo_runner",
     "combo_runner_to_ds",
     "case_runner",
     "find_union_coords",
     "all_missing_ds",
     "case_runner_to_ds",
+    "case_runner_to_df",
     "find_missing_cases",
     "fill_missing_cases",
     "Crop",
@@ -111,6 +119,8 @@ __all__ = [
     "cache_to_disk",
     "save_ds",
     "load_ds",
+    "save_df",
+    "load_df",
     "trimna",
     "sort_dims",
     "check_runs",